+++ src/Tecan/Rack/MPCDNA.php
@@ -2,7 +2,7 @@
 
 namespace MLL\Utils\Tecan\Rack;
 
-use MLL\Utils\Microplate\CoordinateSystem96Well;
+use MLL\Utils\Microplate\CoordinateSystem12x8;
 
 /**
  * @template TContent
@@ -13,7 +13,7 @@ class MPCDNA extends BaseRack
 {
     public function __construct()
     {
-        parent::__construct(new CoordinateSystem96Well());
+        parent::__construct(new CoordinateSystem12x8());
     }
 
     public function type(): string