+++ tests/Microplate/SectionedMicroplate/SectionTest.php
@@ -2,7 +2,7 @@
 
 namespace MLL\Utils\Tests\Microplate\SectionedMicroplate;
 
-use MLL\Utils\Microplate\CoordinateSystem96Well;
+use MLL\Utils\Microplate\CoordinateSystem12x8;
 use MLL\Utils\Microplate\Exceptions\MicroplateIsFullException;
 use MLL\Utils\Microplate\Section;
 use MLL\Utils\Microplate\SectionedMicroplate;
@@ -12,7 +12,7 @@ final class SectionTest extends TestCase
 {
     public function testSectionThrowsWhenFull(): void
     {
-        $coordinateSystem = new CoordinateSystem96Well();
+        $coordinateSystem = new CoordinateSystem12x8();
         $sectionedMicroplate = new SectionedMicroplate($coordinateSystem);
         self::assertCount(0, $sectionedMicroplate->sections);